pom project / unit tests / fixes

+++ src/main/java/com/github/KluevJakov/client/Client.java
@@ -1,5 +1,5 @@
 package com.github.KluevJakov.client;
 
 public interface Client {
-    int paymentLimit();
+    double paymentLimit();
 }
